Fix #330 no more calibration corners sticking to cursor
