+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Warning no longer flashes when dragging on Android, Chromebook, and in Firefox
 - Full screen icon enter/exit swapped to be correct versions
 - Calibration is saved on keyboard input (not just mouse click)
+- Calibration corners no longer stick to cursor on mouse up
 
 ### Removed
