+++ app/[locale]/calibrate/page.tsx
@@ -44,6 +44,7 @@ import pointsReducer from "@/_reducers/pointsReducer";
 import Filters from "@/_components/filters";
 import CalibrationContext, {
   getCalibrationContext,
+  getCalibrationContextUpdatedWithEvent,
   getIsInvalidatedCalibrationContext,
   getIsInvalidatedCalibrationContextWithPointerEvent,
   logCalibrationContextDifferences,
@@ -63,6 +64,7 @@ import useLayers from "@/_hooks/use-layers";
 import ExpandMoreIcon from "@/_icons/expand-more-icon";
 import { LoadStatusEnum } from "@/_lib/load-status-enum";
 import LoadingSpinner from "@/_icons/loading-spinner";
+import { Point } from "@/_lib/point";
 
 const defaultStitchSettings = {
   columnCount: 1,
@@ -128,6 +130,7 @@ export default function Page() {
   const [showCalibrationAlert, setShowCalibrationAlert] = useState(false);
   const [fullScreenTooltipVisible, setFullScreenTooltipVisible] =
     useState(true);
+  const [dragPoint, setDragPoint] = useState<Point | null>(null);
 
   const timeoutRef = useRef<NodeJS.Timeout | null>(null);
 
@@ -357,7 +360,25 @@ export default function Page() {
     setMenusHidden(false);
   }
 
+  function handlePointerUp(e: React.PointerEvent) {
+    /* Nothing to do. This short circuit is required to prevent setting
+     * the localStorage of the points to invalid values */
+    if (dragPoint === null) return;
+
+    localStorage.setItem(
+      "calibrationContext",
+      JSON.stringify(
+        getCalibrationContextUpdatedWithEvent(e, fullScreenHandle.active),
+      ),
+    );
+    dispatch({ type: "set", points });
+    setDragPoint(null);
+  }
   function handlePointerMove(e: React.PointerEvent) {
+    if (e.buttons === 0 && dragPoint !== null) {
+      handlePointerUp(e);
+    }
+
     // Chromebook triggers move after menu hides #268
     if (e.movementX === 0 && e.movementY === 0) {
       return;
@@ -408,6 +429,7 @@ export default function Page() {
   return (
     <main
       onPointerDown={handlePointerDown}
+      onPointerUp={handlePointerUp}
       onPointerMove={handlePointerMove}
       ref={noZoomRefCallback}
       className={`${menusHidden && "cursor-none"} ${isDarkTheme(displaySettings.theme) && "dark bg-black"} w-screen h-screen absolute overflow-hidden touch-none`}
@@ -459,6 +481,8 @@ export default function Page() {
               corners={corners}
               setCorners={setCorners}
               fullScreenHandle={fullScreenHandle}
+              dragPoint={dragPoint}
+              setDragPoint={setDragPoint}
             />
           )}
           {isCalibrating && showingMovePad && (
